backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.6.0
pydantic==2.9.2
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.12
python-dotenv==1.0.1
